move piped.go
